some fail-fast asserts about merge files  (#678)

+++ state/merge.go
@@ -214,6 +214,11 @@ func (d *Domain) staticFilesInRange(r DomainRanges) (valuesFiles, indexFiles, hi
 			valuesFiles = append(valuesFiles, item)
 			return true
 		})
+		for _, f := range valuesFiles {
+			if f == nil {
+				panic("must not happen")
+			}
+		}
 	}
 	return
 }
@@ -232,6 +237,12 @@ func (ii *InvertedIndex) staticFilesInRange(startTxNum, endTxNum uint64) ([]*fil
 		files = append(files, item)
 		return true
 	})
+	for _, f := range files {
+		if f == nil {
+			panic("must not happen")
+		}
+	}
+
 	return files, startJ
 }
 
@@ -252,6 +263,12 @@ func (h *History) staticFilesInRange(r HistoryRanges) (indexFiles, historyFiles 
 			historyFiles = append(historyFiles, item)
 			return true
 		})
+
+		for _, f := range historyFiles {
+			if f == nil {
+				panic("must not happen")
+			}
+		}
 	}
 	return
 }
@@ -776,9 +793,15 @@ func (h *History) mergeFiles(indexFiles, historyFiles []*filesItem, r HistoryRan
 }
 
 func (d *Domain) integrateMergedFiles(valuesOuts, indexOuts, historyOuts []*filesItem, valuesIn, indexIn, historyIn *filesItem) {
+	if valuesIn == nil {
+		panic("must not happen")
+	}
 	d.History.integrateMergedFiles(indexOuts, historyOuts, indexIn, historyIn)
 	d.files.ReplaceOrInsert(valuesIn)
 	for _, out := range valuesOuts {
+		if out == nil {
+			panic("must not happen")
+		}
 		d.files.Delete(out)
 		out.decompressor.Close()
 		out.index.Close()
@@ -786,8 +809,14 @@ func (d *Domain) integrateMergedFiles(valuesOuts, indexOuts, historyOuts []*file
 }
 
 func (ii *InvertedIndex) integrateMergedFiles(outs []*filesItem, in *filesItem) {
+	if in == nil {
+		panic("must not happen")
+	}
 	ii.files.ReplaceOrInsert(in)
 	for _, out := range outs {
+		if out == nil {
+			panic("must not happen")
+		}
 		ii.files.Delete(out)
 		out.decompressor.Close()
 		out.index.Close()
@@ -795,9 +824,15 @@ func (ii *InvertedIndex) integrateMergedFiles(outs []*filesItem, in *filesItem) 
 }
 
 func (h *History) integrateMergedFiles(indexOuts, historyOuts []*filesItem, indexIn, historyIn *filesItem) {
+	if historyIn == nil {
+		panic("must not happen")
+	}
 	h.InvertedIndex.integrateMergedFiles(indexOuts, indexIn)
 	h.files.ReplaceOrInsert(historyIn)
 	for _, out := range historyOuts {
+		if out == nil {
+			panic("must not happen")
+		}
 		h.files.Delete(out)
 		out.decompressor.Close()
 		out.index.Close()
